Add shuffle_playlist.ipynb
This notebook contains a Python script to shuffle the custom order of your Spotify playlist. By default, Spotify maintains a consistent ordering of tracks in a playlist, which can result in a predictable listening experience. This notebook solves this experience by providing a quick solution that reorders your tracks randomly.